requirements.txt
autopep8==1.6.0
cycler==0.11.0
fonttools==4.31.2
greenlet==1.1.2
kiwisolver==1.4.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.32
toml==0.10.2
